--- js/Store.js
@@ -1,16 +1,0 @@
-class Store {
-  #pageNumber;
-  constructor() {
-    this.#pageNumber = 0;
-    this.setPage = function (pagenumber) {
-      this.#pageNumber = pagenumber;
-    };
-    this.getPage = function () {
-      return this.#pageNumber;
-    };
-  }
-}
-
-const Stores = new Store();
-
-export default Stores;
