+++ js/render/renderGrid.js
@@ -5,37 +5,35 @@ import {
   listMain,
 } from "./renderMain.js";
 import { shuffle } from "../../utils/utils.js";
-import { progressInterval } from "../category.js";
 import logo from "../../json/news_image.json" assert { type: "json" };
-import Stores from "../Store.js";
+import Stores from "../../utils/Store.js";
 
 const MAX_PAGE_NUMBER = 3;
 const MIN_PAGE_NUMBER = 0;
 let logos = logo;
 
 function renderGrid() {
-  console.log(Stores.getPage());
   const COUNT_PER_PAGE = 24;
-  if (clearInterval != undefined) clearInterval(progressInterval);
   shuffle(logos);
   gridMain.style.display = "grid";
   listMain.style.display = "none";
   gridMain.innerHTML = "";
+  let outerDiv = "";
   for (
     let LOGO_INDEX = Stores.getPage() * COUNT_PER_PAGE;
     LOGO_INDEX < COUNT_PER_PAGE * Stores.getPage() + 24;
     LOGO_INDEX++
   ) {
-    const outerDiv = document.createElement("div");
-    outerDiv.append(drawLogo(LOGO_INDEX));
-    gridMain.append(outerDiv);
+    outerDiv += `<div class="grid-list"><div class="hover-subscribe-button"><img src="./img/subscribe_button.svg"></div>${drawLogo(
+      LOGO_INDEX
+    )}</div>`;
   }
+  gridMain.innerHTML = outerDiv;
 }
 
 function drawLogo(LOGO_INDEX) {
   if (logos[LOGO_INDEX] != undefined) {
-    const newsLogo = document.createElement("img");
-    newsLogo.src = `${logos[LOGO_INDEX].logo}`;
+    const newsLogo = `<img class="grid-image" src="${logos[LOGO_INDEX].logo}">`;
     return newsLogo;
   }
   return "";
